app/requirements.txt
torch
torchvision
transformers
numpy
Click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
gunicorn==20.0.4
psycopg2==2.9.1
Pillow
flask
requests
flask-sqlalchemy